Revert "Feat 1031 add theme for n12"
