+++ src/App.tsx
@@ -11,7 +11,6 @@ import { useTranslation } from 'react-i18next';
 import { useTheme } from '@material-ui/core/styles';
 import PopUpRedirect from './components/atoms/PopUpRedirect';
 import WidgetsTemplate from './components/organisms/WidgetsTemplate';
-import {observer} from "mobx-react-lite";
 // main components height - must add up to 100
 
 const headerHeight = '5vh';
@@ -71,4 +70,4 @@ const App: FC = () => {
     </StoreContext.Provider>
   );
 };
-export default observer(App);
+export default App;
