+++ src/style/_colors.ts
@@ -1,20 +1,34 @@
-import {barColors, blueVioletColor, pieChartColors, silverGrayColor} from "./default/_defaultColors";
-import {
-  grayBeige,
-  mainN12Red,
-  n12BackgroundColor,
-  n12BarColors,
-  n12DataContainerColor,
-  n12PieChartColors
-} from "./n12/_n12_colors";
-
 export const blackColor = '#000';
 export const whiteColor = '#ffffff';
 export const transparentColor = 'transparent';
+//blue base collection
+export const blueVioletColor = '#3f51b5';
+export const skyBlueColor = '#00aaff';
+export const oceanBlueColor = '#0088cc';
+export const secondaryBgColor = '#E3F6FF';
+//white-Gray base collection
+export const shadowColor = '#4c4c4c';
+export const brightGreyColor = '#5a5b5e';
+export const smokeWhiteColor = '#f5f5f5';
+//silver-Gray base collection
+export const silverSmokeColor = '#D3D3D3';
+export const silverGrayColor = '#A2A2A2';
+export const silverSpoonColor = '#647171';
+//red base collection
+export const roseColor = '#c43a31';
+export const cherryJamColor = '#8a1212';
+export const honeyColor = '#F2BB0A';
+//Stacked chart
+export const yellowColor = '#ffdd22';
+// Grey
+export const darkGrey = '#818386';
 
 // Opacity
 export const opacity80percent = '8c';
 export const transparent = 'transparent';
+
+//pie
+export const pieChartColors = ['#AE0721', '#818386', '#d90000', '#890505', '#6a6a6a'];
 
 //road number icon && font
 export const roadIconColors = {
@@ -25,45 +39,3 @@ export const roadIconColors = {
 };
 
 export const tooltipMarkerBorderColorArrow = 'transparent #000000 transparent #000000';
-
-
-declare module "@material-ui/core";
-interface PaletteColorOptions {
-  light?: string;
-  main: string;
-  dark?: string;
-  contrastText?: string;
-}
-
-export interface ColorScheme extends PaletteColorOptions {
-  fontColor: string;
-  backgroundColor: string;
-  containerColor: string;
-  titleContainerColor: string;
-  pieChartColors: string[];
-  barChartColors: string[];
-  roadNumberBackground?: string;
-}
-
-export const defaultThemeColors : ColorScheme = {
-  main: blueVioletColor,
-  fontColor: blackColor,
-  backgroundColor: whiteColor,
-  containerColor: whiteColor,
-  titleContainerColor: whiteColor,
-  pieChartColors: pieChartColors,
-  barChartColors: barColors,
-  roadNumberBackground: silverGrayColor,
-}
-
-export const n12ThemeColors : ColorScheme = {
-  main: mainN12Red,
-  fontColor: whiteColor,
-  backgroundColor: n12BackgroundColor,
-  containerColor: n12DataContainerColor,
-  titleContainerColor: mainN12Red,
-  pieChartColors: n12PieChartColors,
-  barChartColors: n12BarColors,
-  roadNumberBackground: grayBeige,
-}
-
