+++ src/style/_fonts.ts
@@ -8,14 +8,3 @@ export const fontFamilyString = [
   'Arial',
   'sans-serif',
 ].join(',');
-
-export const n12FontFamilyString = [
-  'Rubik',
-  '-apple-system',
-  'BlinkMacSystemFont',
-  '"Segoe UI"',
-  'Roboto',
-  '"Helvetica Neue"',
-  'Arial',
-  'sans-serif',
-].join(',');
