+++ src/style/index.ts
@@ -1,6 +1,4 @@
 export * from './_fonts';
 export * from './_colors';
-export * from './default/_defaultColors';
 export * from './_sizes';
-export { defaultThemeOptions } from './default/defaultTheme';
-export { n12ThemeOptions } from './n12/n12Theme';
+export { defaultThemeOptions } from './theme';
